+++ src-tauri/src/minecraft/utils/mod.rs
@@ -1,3 +1,4 @@
 //! Utils module
 
 pub mod file_checker;
+pub mod maven;